One plus One solver works, started to create solver wrapper class

+++ GeneticAlgorithms/Genome.h
@@ -10,5 +10,5 @@ struct Genome
 {
 	T data[size];
 	size_t length = size;
-	int32_t fitness;
+	int64_t fitness;
 };
